# backend/docs/swagger.yaml
basePath: /
definitions:
...
  consts.SourceType:
    enum:
    - dingtalk
    - feishu
    - wecom
    - oauth
    - github
    - cas
    - ldap
    - widget
    - dingtalk_bot
    - feishu_bot
    - lark_bot
    - wechat_bot
    - wecom_ai_bot
    - wechat_service_bot
    - discord_bot
    - wechat_official_account
    - openai_api
    - mcp_server
    type: string
    x-enum-varnames:
    - SourceTypeDingTalk
    - SourceTypeFeishu
    - SourceTypeWeCom
    - SourceTypeOAuth
    - SourceTypeGitHub
    - SourceTypeCAS
    - SourceTypeLDAP
    - SourceTypeWidget
    - SourceTypeDingtalkBot
    - SourceTypeFeishuBot
    - SourceTypeLarkBot
    - SourceTypeWechatBot
    - SourceTypeWecomAIBot
    - SourceTypeWechatServiceBot
    - SourceTypeDiscordBot
    - SourceTypeWechatOfficialAccount
    - SourceTypeOpenAIAPI
    - SourceTypeMcpServer
...
  consts.UserKBPermission:
    enum:
    - ""
    - not null
    - full_control
    - doc_manage
    - data_operate
    type: string
    x-enum-comments:
      UserKBPermissionDataOperate: 数据运营
      UserKBPermissionDocManage: 文档管理
      UserKBPermissionFullControl: 完全控制
      UserKBPermissionNotNull: 有权限
      UserKBPermissionNull: 无权限
    x-enum-descriptions:
    - 无权限
    - 有权限
    - 完全控制
    - 文档管理
    - 数据运营
    x-enum-varnames:
    - UserKBPermissionNull
    - UserKBPermissionNotNull
    - UserKBPermissionFullControl
    - UserKBPermissionDocManage
    - UserKBPermissionDataOperate
  consts.UserRole:
    enum:
    - admin
    - user
    type: string
    x-enum-comments:
      UserRoleAdmin: 管理员
      UserRoleUser: 普通用户
    x-enum-descriptions:
    - 管理员
    - 普通用户
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleUser
...
  domain.AccessSettings:
    properties:
      base_url:
        type: string
      enterprise_auth:
        $ref: '#/definitions/domain.EnterpriseAuth'
      hosts:
        items:
          type: string
        type: array
      is_forbidden:
        description: 禁止访问
        type: boolean
      ports:
        items:
          type: integer
        type: array
      private_key:
        type: string
      public_key:
        type: string
      simple_auth:
        $ref: '#/definitions/domain.SimpleAuth'
      source_type:
        allOf:
        - $ref: '#/definitions/consts.SourceType'
        description: 企业认证来源
      ssl_ports:
        items:
          type: integer
        type: array
      trusted_proxies:
        items:
          type: string
        type: array
    type: object
...
  domain.CreateKBReleaseReq:
    properties:
      kb_id:
        type: string
      message:
        type: string
      node_ids:
        description: create release after these nodes published
        items:
          type: string
        type: array
      tag:
        type: string
    required:
    - kb_id
    - message
    - tag
    type: object
  domain.CreateKnowledgeBaseReq:
    properties:
      hosts:
        items:
          type: string
        type: array
      name:
        type: string
      ports:
        items:
          type: integer
        type: array
      private_key:
        type: string
      public_key:
        type: string
      ssl_ports:
        items:
          type: integer
        type: array
    required:
    - name
    type: object
...
  domain.EnterpriseAuth:
    properties:
      enabled:
        type: boolean
    type: object
...
  domain.GetKBReleaseListResp:
    properties:
      data:
        items:
          $ref: '#/definitions/domain.KBReleaseListItemResp'
        type: array
      total:
        type: integer
    type: object
...
  domain.KBReleaseListItemResp:
    properties:
      created_at:
        type: string
      id:
        type: string
      kb_id:
        type: string
      message:
        type: string
      publisher_account:
        type: string
      tag:
        type: string
    type: object
  domain.KnowledgeBaseDetail:
    properties:
      access_settings:
        $ref: '#/definitions/domain.AccessSettings'
      created_at:
        type: string
      dataset_id:
        type: string
      id:
        type: string
      name:
        type: string
      perm:
        allOf:
        - $ref: '#/definitions/consts.UserKBPermission'
        description: 用户对知识库的权限
      updated_at:
        type: string
    type: object
...
  domain.PWResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  domain.SimpleAuth:
    properties:
      enabled:
        type: boolean
      password:
        type: string
    type: object
...
  domain.UpdateKnowledgeBaseReq:
    properties:
      access_settings:
        $ref: '#/definitions/domain.AccessSettings'
      id:
        type: string
      name:
        type: string
    required:
    - id
    type: object
...
  v1.CreateUserReq:
    properties:
      account:
        type: string
      password:
        minLength: 8
        type: string
      role:
        allOf:
        - $ref: '#/definitions/consts.UserRole'
        enum:
        - admin
        - user
    required:
    - account
    - password
    - role
    type: object
  v1.CreateUserResp:
    properties:
      id:
        type: string
    type: object
...
info:
  contact: {}
  description: panda-wiki API documentation
  title: panda-wiki API
  version: "1.0"
paths:
...
  /api/v1/knowledge_base:
    post:
      consumes:
      - application/json
      description: CreateKnowledgeBase
      parameters:
      - description: CreateKnowledgeBase Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateKnowledgeBaseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: CreateKnowledgeBase
      tags:
      - knowledge_base
  /api/v1/knowledge_base/detail:
    delete:
      consumes:
      - application/json
      description: DeleteKnowledgeBase
      parameters:
      - description: Knowledge Base ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: DeleteKnowledgeBase
      tags:
      - knowledge_base
    get:
      consumes:
      - application/json
      description: GetKnowledgeBaseDetail
      parameters:
      - description: Knowledge Base ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.KnowledgeBaseDetail'
              type: object
      security:
      - bearerAuth: []
      summary: GetKnowledgeBaseDetail
      tags:
      - knowledge_base
    put:
      consumes:
      - application/json
      description: UpdateKnowledgeBase
      parameters:
      - description: UpdateKnowledgeBase Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateKnowledgeBaseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: UpdateKnowledgeBase
      tags:
      - knowledge_base
...
  /api/v1/knowledge_base/release:
    post:
      consumes:
      - application/json
      description: CreateKBRelease
      parameters:
      - description: CreateKBRelease Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateKBReleaseReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: CreateKBRelease
      tags:
      - knowledge_base
  /api/v1/knowledge_base/release/list:
    get:
      consumes:
      - application/json
      description: GetKBReleaseList
      parameters:
      - description: Knowledge Base ID
        in: query
        name: kb_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.PWResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.GetKBReleaseListResp'
              type: object
      summary: GetKBReleaseList
      tags:
      - knowledge_base
...
  /api/v1/user/create:
    post:
      consumes:
      - application/json
      description: CreateUser
      parameters:
      - description: CreateUser Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.CreateUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/domain.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.CreateUserResp'
              type: object
      summary: CreateUser
      tags:
      - user
  /api/v1/user/delete:
    delete:
      consumes:
      - application/json
      description: DeleteUser
      parameters:
      - in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
      summary: DeleteUser
      tags:
      - user
...
securityDefinitions:
  bearerAuth:
    description: Type "Bearer" + a space + your token to authorize
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"